Refactoring HistoryDialog on ViewVersionDialog and CompareVersionsDialog(WIP) HistoryDialog.tsx: moving VersionList, creating Pagination Wrapper;
