+++ ui/app/src/state/reducers/dialogs/history.ts
@@ -17,10 +17,9 @@
 import { createAction, createReducer } from '@reduxjs/toolkit';
 import StandardAction from '../../../models/StandardAction';
 import GlobalState from '../../../models/GlobalState';
-import { Compare, HistoryDialogStateProps } from '../../../modules/Content/History/HistoryDialog';
+import { HistoryDialogStateProps } from '../../../modules/Content/History/HistoryDialog';
 import { VersionsResponse } from '../../../models/version';
 import { AjaxError } from 'rxjs/ajax';
-import { createEntityState, createLookupTable } from '../../../utils/object';
 
 interface HistoryConfigProps {
   path: string;
@@ -41,8 +40,6 @@ export const fetchItemVersionsComplete = createAction<VersionsResponse>('FETCH_I
 
 export const fetchItemVersionsFailed = createAction<AjaxError>('FETCH_ITEM_VERSIONS_FAILED');
 
-export const compareHistories = createAction<Compare>('COMPARE_HISTORIES');
-
 export const historyDialogChangePage = createAction<number>('HISTORY_DIALOG_CHANGE_PAGE');
 
 export const revertContent = createAction<string>('REVERT_CONTENT');
@@ -60,13 +57,9 @@ const initialState = {
   path: null,
   config: null,
   environment: null,
-  module: null,
-  compare: {
-    a: null,
-    b: null
-  },
-  order: null,
-  ...createEntityState()
+  error: null,
+  isFetching: null,
+  versions: null,
 };
 
 export default createReducer<GlobalState['dialogs']['history']>(
@@ -89,20 +82,14 @@ export default createReducer<GlobalState['dialogs']['history']>(
     }),
     [fetchItemVersionsComplete.type]: (state, { payload }) => ({
       ...state,
-      byId: createLookupTable(payload.versions, 'versionNumber'),
+      versions: payload.versions,
       current: payload.versions[0].versionNumber,
-      order: payload.versions.map(version => version.versionNumber),
-      item: payload.item,
       isFetching: false
     }),
     [fetchItemVersionsFailed.type]: (state, { payload }) => ({
       ...state,
       error: payload.response,
       isFetching: false
-    }),
-    [compareHistories.type]: (state, { payload }) => ({
-      ...state,
-      compare: { ...state.compare, ...payload }
     }),
     [historyDialogChangePage.type]: (state, { payload }) => ({
       ...state,
